+++ src/main/java/kitchenpos/products/tobe/domain/DisplayedName.java
@@ -1,6 +1,5 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.exception.InvalidProductNameException;
 
 import javax.persistence.Column;
